Added basic opcode decoding
Added opcode combining (taking two single bytes and combining them into
a single opcode number)

This is done by taking the first byte in memory, and shifting it 8 bits
to the left. This creates a 2 byte number where the lower 8 bits are
all zeroes. Then you OR that 2-byte number with the next byte in memory.
Since all the lower end numbers are 0, any 1 in the second single-byte
number will be transferred to the result through the OR operation.